Add enabledReplacementsRegex and customReplacementsRegex settings
