+++ src/replacement/replacementMax.ts
@@ -1,0 +1,377 @@
+import { ReplacementCategory } from './replacementTypes';
+
+// Maximum style replacements for LaTeX; by default not enabled in package.json
+export const MAX_STYLE_REPLACEMENTS: ReplacementCategory = {
+  name: 'max_style',
+  description: 'Maximum style replacements for LaTeX commands and symbols',
+  isRegex: false,
+  patterns: {
+    // Basic Math Operators
+    '\\text{tr}': '\\tr',
+    '\\text{Tr}': '\\tr',
+    // '\\Tr': '\\tr',
+    '\\text{sign}': '\\sign',
+    '\\text{argmin}': '\\argmin',
+    '\\text{argmax}': '\\argmax',
+    '\\text{sort}': '\\sort',
+    '\\text{argsort}': '\\argsort',
+    '\\text{Cat}': '\\Cat',
+    '\\text{Bern}': '\\Bern',
+    '\\text{Unif}': '\\Unif',
+    '\\marhrm{const}': '\\const',
+    '\\text{const}': '\\const',
+    '\\text{det}': '\\det',
+    '\\text{tot}': '\\tot',
+    '_{tot}': '_{\\tot}',
+    '^{tot}': '^{\\tot}',
+    '^\\intercal': '^\\top',
+    '^{\\intercal}': '^\\top',
+
+    // Common Math Shortcuts
+    '\\frac{1}{2}': '\\ha',
+    '\\frac{1}{\\sqrt{2}}': '\\sha',
+    '\\mathds{1}': '\\Id',
+    '\\boldsymbol{0}': '\\bzero',
+    '\\boldsymbol{1}': '\\bone',
+    '\\mathrm{d}': '\\dd',
+
+    // Greek Letters (Regular)
+    '\\alpha': '\\al',
+    '\\beta': '\\bt',
+    '\\gamma': '\\ga',
+    '\\delta': '\\de',
+    '\\varepsilon': '\\eps',
+    '\\theta': '\\ta',
+    '\\Theta': '\\Ta',
+    '\\kappa': '\\ka',
+    '\\lambda': '\\la',
+    '\\omega': '\\om',
+    '\\Omega': '\\Om',
+    '\\sigma': '\\sg',
+    '\\Sigma': '\\Sig',
+    '\\varphi': '\\vphi',
+
+    // Correct spelling errors of greek letters
+    '\\\\bet': '\\bet',
+    '\\\\bze': '\\bze',
+    '\\\\cP': '\\cP',
+    '\\\\Om': '\\Om',
+
+    // Greek Letters (Bold)
+    '\\boldsymbol{\\alpha}': '\\bal',
+    '\\boldsymbol{\\beta}': '\\bbt',
+    '\\boldsymbol{\\chi}': '\\bch',
+    '\\boldsymbol{\\epsilon}': '\\beps',
+    '\\boldsymbol{\\eta}': '\\bet',
+    '\\boldsymbol{\\gamma}': '\\bga',
+    '\\boldsymbol{\\mu}': '\\bmu',
+    '\\boldsymbol{\\nu}': '\\bnu',
+    '\\boldsymbol{\\omega}': '\\bom',
+    '\\boldsymbol{\\phi}': '\\bphi',
+    '\\boldsymbol{\\pi}': '\\bpi',
+    '\\boldsymbol{\\sigma}': '\\bsigma',
+    '\\boldsymbol{\\theta}': '\\bta',
+    '\\boldsymbol{\\varphi}': '\\bvphi',
+    '\\boldsymbol{\\xi}': '\\bxi',
+    '\\boldsymbol{\\zeta}': '\\bze',
+    '\\boldsymbol{\\Sigma}': '\\bSig',
+    '\\boldsymbol{\\lambda}': '\\bla',
+    '\\boldsymbol{\\Gamma}': '\\bGa',
+    '\\boldsymbol{\\Lambda}': '\\bLa',
+    '\\bSigma': '\\bSig',
+
+    // Hat Variables
+    '\\hat{\\sigma}': '\\hsg',
+    '\\hat{\\Sigma}': '\\hSig',
+    '\\hat{\\pi}': '\\hpi',
+    '\\hat{\\rho}': '\\hrho',
+    '\\hat{H}': '\\hH',
+    '\\hat{F}': '\\hF',
+    '\\hat{P}': '\\hP',
+    '\\hat{\\mathbf{n}}': '\\hbn',
+    '\\hat{\\mathbf{v}}': '\\hbv',
+    '\\hat{\\boldsymbol{\\zeta}}': '\\hbze',
+
+    '\\frac12': '\\ha',
+
+    // Mathcal Letters
+    '\\mathcal{A}': '\\cA',
+    '\\mathcal{B}': '\\cB',
+    '\\mathcal{C}': '\\cC',
+    '\\mathcal{D}': '\\cD',
+    '\\mathcal{E}': '\\cE',
+    '\\mathcal{F}': '\\cF',
+    '\\mathcal{G}': '\\cG',
+    '\\mathcal{H}': '\\cH',
+    '\\mathcal{I}': '\\cI',
+    '\\mathcal{J}': '\\cJ',
+    '\\mathcal{K}': '\\cK',
+    '\\mathcal{L}': '\\cL',
+    '\\mathcal{M}': '\\cM',
+    '\\mathcal{N}': '\\cN',
+    '\\mathcal{O}': '\\cO',
+    '\\mathcal{P}': '\\cP',
+    '\\mathcal{Q}': '\\cQ',
+    '\\mathcal{R}': '\\cR',
+    '\\mathcal{S}': '\\cS',
+    '\\mathcal{T}': '\\cT',
+    '\\mathcal{U}': '\\cU',
+    '\\mathcal{V}': '\\cV',
+    '\\mathcal{W}': '\\cW',
+    '\\mathcal{X}': '\\cX',
+    '\\mathcal{Y}': '\\cY',
+    '\\mathcal{Z}': '\\cZ',
+
+    // Mathbb Letters
+    '\\mathbb{C}': '\\eC',
+    '\\mathbb{E}': '\\eE',
+    '\\mathbb{I}': '\\eI',
+    '\\mathbb{N}': '\\eN',
+    '\\mathbb{P}': '\\eP',
+    '\\mathbb{Q}': '\\eQ',
+    '\\mathbb{R}': '\\eR',
+    '\\mathbb{T}': '\\eT',
+    '\\mathbb{V}': '\\eV',
+
+    // Mathbf Letters (Lowercase)
+    '\\mathbf{a}': '\\ba',
+    '\\mathbf{b}': '\\bb',
+    '\\mathbf{c}': '\\bc',
+    '\\mathbf{e}': '\\be',
+    '\\mathbf{f}': '\\bbf',
+    '\\mathbf{g}': '\\bg',
+    '\\mathbf{h}': '\\bh',
+    '\\mathbf{j}': '\\bj',
+    '\\mathbf{n}': '\\bn',
+    '\\mathbf{p}': '\\bp',
+    '\\mathbf{q}': '\\bq',
+    '\\mathbf{r}': '\\br',
+    '\\mathbf{s}': '\\bs',
+    '\\mathbf{u}': '\\bu',
+    '\\mathbf{v}': '\\bv',
+    '\\mathbf{w}': '\\bw',
+    '\\mathbf{x}': '\\bx',
+    '\\mathbf{y}': '\\by',
+    '\\mathbf{z}': '\\bz',
+
+    // Mathbf Letters (Uppercase)
+    '\\mathbf{A}': '\\bA',
+    '\\mathbf{B}': '\\bB',
+    '\\mathbf{C}': '\\bC',
+    '\\mathbf{D}': '\\bD',
+    '\\mathbf{E}': '\\bE',
+    '\\mathbf{F}': '\\bF',
+    '\\mathbf{G}': '\\bG',
+    '\\mathbf{I}': '\\bI',
+    '\\mathbf{J}': '\\bJ',
+    '\\mathbf{K}': '\\bK',
+    '\\mathbf{M}': '\\bM',
+    '\\mathbf{Q}': '\\bQ',
+    '\\mathbf{R}': '\\bR',
+    '\\mathbf{U}': '\\bU',
+    '\\mathbf{V}': '\\bV',
+    '\\mathbf{W}': '\\bW',
+    '\\mathbf{X}': '\\bX',
+    '\\mathbf{Y}': '\\bY',
+    '\\mathbf{Z}': '\\bZ',
+
+    // Bar Variables
+    '\\bar{\\rho}': '\\barrho',
+    '\\bar{H}': '\\barH',
+    '\\bar{S}': '\\barS',
+    '\\bar{\\alpha}': '\\baral',
+    '\\bar{\\mathbf{v}}': '\\barbv',
+
+    // Tilde Variables
+    '\\tilde{\\mathcal{H}}': '\\tcH',
+    '\\tilde{\\mathcal{Q}}': '\\tcQ',
+    '\\tilde{\\mathcal{S}}': '\\tcS',
+    '\\tilde{\\mathcal{W}}': '\\tcW',
+    '\\tilde{\\gamma}': '\\tga',
+    '\\tilde{\\lambda}': '\\tla',
+    '\\tilde{\\phi}': '\\tphi',
+    '\\tilde{\\psi}': '\\tpsi',
+    '\\tilde{\\rho}': '\\trho',
+    '\\tilde{\\Sigma}': '\\tSig',
+    '\\tilde{\\mu}': '\\tmu',
+    '\\tilde{\\tau}': '\\ttau',
+    '\\tilde{\\mathbf{a}}': '\\tba',
+    '\\tilde{\\mathbf{x}}': '\\tbx',
+    '\\tilde{\\mathbf{v}}': '\\tbv',
+    '\\tilde{\\mathbf{p}}': '\\tbp',
+    '\\tilde{\\mathbf{w}}': '\\tbw',
+    '\\tilde{\\mathbf{z}}': '\\tbz',
+    '\\tilde{\\mathbf{M}}': '\\tbM',
+    '\\tilde{\\mathbf{T}}': '\\tbT',
+    '\\tilde{\\mathbf{X}}': '\\tbX',
+    '\\tilde{\\boldsymbol{\\zeta}}': '\\tbze',
+    '\\tilde{\\boldsymbol{\\gamma}}': '\\tbga',
+    '\\tilde{\\boldsymbol{\\lambda}}': '\\tbla',
+    '\\tilde{\\boldsymbol{\\pi}}': '\\tbpi',
+    '\\tilde{\\boldsymbol{\\xi}}': '\\tbxi',
+    '\\tilde{\\boldsymbol{\\eta}}': '\\tbet',
+    '\\tilde{\\boldsymbol{\\Gamma}}': '\\tbGa',
+    '\\tilde{0}': '\\tzero',
+    '\\tilde{1}': '\\tone',
+    '\\tilde{t}': '\\tit',
+    '\\tilde{f}': '\\tif',
+    '\\tilde{x}': '\\tx',
+    '\\tilde{z}': '\\tz',
+    '\\tilde{p}': '\\tp',
+    '\\tilde{q}': '\\tq',
+    '\\tilde{B}': '\\tB',
+    '\\tilde{F}': '\\tF',
+    '\\tilde{J}': '\\tJ',
+    '\\tilde{M}': '\\tM',
+    '\\tilde{P}': '\\tP',
+    '\\tilde{T}': '\\tT',
+    '\\tilde{Z}': '\\tZ',
+
+    // Vector Variables
+    '\\vec{p}': '\\vp',
+    '\\vec{q}': '\\vq',
+    '\\vec{v}': '\\vv',
+    '\\vec{x}': '\\vx',
+    '\\vec{y}': '\\vy',
+
+    // Physics and Statistical Mechanics
+    'H^{\\text{eff}}': '\\effH',
+    '\\mathcal{H}^{\\text{eff}}': '\\ceffH',
+    'p^{\\text{eq}}': '\\peq',
+    'q^{\\text{eq}}': '\\qeq',
+    '\\rho^{\\text{eq}}': '\\rhoeq',
+    '\\rho^{\\text{st}}': '\\rhost',
+
+    // Arrows and Relations
+    '\\rightarrow': '\\ra',
+    '\\leftarrow': '\\lar',
+    '\\leftrightarrow': '\\lra',
+    '\\Leftarrow': '\\La',
+    '\\Rightarrow': '\\Ra',
+    '\\Leftrightarrow': '\\LRa',
+    '\\not\\implies': '\\nimplies',
+
+    // Calculus
+    '\\partial': '\\der',
+    '\\nabla': '\\na',
+    '\\boldsymbol{\\nabla}': '\\bna',
+    '\\text{div}': '\\bdiv',
+    '\\frac{\\partial}{\\partial t}': '\\ddt',
+    '\\frac{\\mathrm{d}}{\\mathrm{d} t}': '\\dddt',
+
+    // Miscellaneous
+    '\\nonumber': '\\nn',
+    '\\dagger': '\\da',
+    '\\backslash': '\\bksl',
+
+    // Differential spacing preferences
+    '\\dd s\\,': '\\dd s~',
+    '\\dd\\bze\\,': '\\dd\\bze~',
+    '\\dd\\bxi\\,': '\\dd\\bxi~',
+    '\\dd\\beta\\,': '\\dd\\beta~',
+    '\\dd\\tau\\,': '\\dd\\tau~',
+    '\\dd\\t\\,': '\\dd\\t~',
+    '\\dd\\bx\\,': '\\dd\\bx~',
+    '\\dd\\bz\\,': '\\dd\\bz~',
+    '\\dd\\bze_{\\tauf}\\,': '\\dd\\bze_{\\tauf}~',
+    '\\dd x ': '\\dd x~ ',
+    '\\dd x': "\\dd x'~'",
+    '\\quad\\Ra': '~~~\\Ra',
+    '\\Ra\,': '\\Ra~',
+    '\\tau_f': '\\tauf',
+    '\\frac{d\\tbx': '\\frac{\\dd\\tbx',
+    '\\frac{d\\bx': '\\frac{\\dd\\bx',
+    '\\frac{d}{d\\tau}': '\\frac{\\dd}{\\dd\\tau}',
+    '\\frac{d}{d\\}': '\\frac{\\dd}{\\dd\\}',
+    '{d t}': '{\\dd t}',
+    '{dt}': '{\\dd t}',
+    '{d x}': '{\\dd x}',
+    '{dx}': '{\\dd x}',
+    '\\frac{d}{d\\': '\\frac{\\dd}{\\dd\\',
+    '{d\\tau}': '{\\dd\\tau}',
+    '{d\\ttau}': '{\\dd\\ttau}',
+    '{d\\bze}': '{\\dd\\bze}',
+    '{d\\bxi}': '{\\dd\\bxi}',
+    '{d\\beta}': '{\\dd\\beta}',
+    '{d\\bx}': '{\\dd\\bx}',
+    '{d\\bz}': '{\\dd\\bz}',
+    ' d\\bz': ' d\\bz~',
+    ' d\\bx': ' d\\bx~',
+    ' d\\bze': ' d\\bze~',
+    ' d\\bxi': ' d\\bxi~',
+    ' d\\beta': ' d\\beta~',
+    ' d\\tau': ' d\\tau~',
+    ' d\\ttau': ' d\\ttau~',
+    '\\int d\\': '\\int \\dd\\',
+
+    '\\int \\dd \\bx \\': '\\int \\dd \\bx~ \\',
+    '\\frac{dS}': '\\frac{\\dd S}',
+    '\\\\ba': '\\ba',
+    '\\\\bb': '\\bb',
+    '\\\\bc': '\\bc',
+    '\\\\be': '\\be',
+    '\\\\bf': '\\bf',
+    '\\\\bg': '\\bg',
+    '\\\\bh': '\\bh',
+    '\\\\bj': '\\bj',
+    '\\\\bn': '\\bn',
+    '\\\\bp': '\\bp',
+    '\\\\bq': '\\bq',
+    '\\\\br': '\\br',
+    '\\\\bs': '\\bs',
+    '\\\\bu': '\\bu',
+    '\\\\bv': '\\bv',
+    '\\\\bw': '\\bw',
+    '\\\\bx': '\\bx',
+    '\\\\by': '\\by',
+    '\\\\bz': '\\bz',
+    '\\\\bA': '\\bA',
+    '\\\\bB': '\\bB',
+    '\\\\bC': '\\bC',
+    '\\\\bD': '\\bD',
+    '\\\\bE': '\\bE',
+    '\\\\bF': '\\bF',
+    '\\\\bG': '\\bG',
+    '\\\\bH': '\\bH',
+    '\\\\bI': '\\bI',
+    '\\\\bJ': '\\bJ',
+    '\\\\bK': '\\bK',
+    '\\\\bL': '\\bL',
+    '\\\\bM': '\\bM',
+    '\\\\bN': '\\bN',
+    '\\\\bP': '\\bP',
+    '\\\\bQ': '\\bQ',
+    '\\\\bR': '\\bR',
+    '\\\\bS': '\\bS',
+    '\\\\bT': '\\bT',
+    '\\\\bU': '\\bU',
+    '\\\\bV': '\\bV',
+    '\\\\bW': '\\bW',
+    '\\\\bX': '\\bX',
+    '\\\\bY': '\\bY',
+    '\\\\bZ': '\\bZ',
+    // equilibrium and steady state
+    '{\\text{ss}}': '{\\text{st}}',
+    '\\rho^{st}': '\\rhost',
+    '\\rho^{eq}': '\\rhoeq',
+    '\\rho^{\\text{ss}}': '\\rhost',
+    '\\rho^{ss}': '\\rhost',
+    '\\rho_{\\text{ss}}': '\\rhost',
+    '\\rho_{ss}': '\\rhost',
+    '\\ln': '\\log',
+    '\\rhost^R': '\\rhost_R',
+    '\\F_': 'F_',
+    '\\F^': 'F^',
+    '\\bepsilon': '\\beps',
+    // '\\mathbf{I}': '\\bI',
+    '{\\rm tot}': '\\text{tot}',
+    '{\\rm eq}': '\\text{eq}',
+    '{\\rm st}': '\\text{st}',
+    '{\\rm na}': '\\text{na}',
+    '\\G(': 'G(',
+    '\\F(': 'F(',
+
+    // Some aggressive ones:
+    '+ O(': '+ \\cO(',
+  },
+};
