+++ src/replacement/replacementTypes.ts
@@ -1,0 +1,7 @@
+export interface ReplacementCategory {
+  name: string;
+  description: string;
+  patterns: { [key: string]: string };
+  isRegex?: boolean;
+  flags?: string; // Optional regex flags
+}
